- Working on using percentage to remember reading position - TODO, create a DB migration

+++ android/src/main/java/app/shosetsu/android/backend/database/DBHelper.kt
@@ -228,7 +228,7 @@ class DBHelper(context: Context) :
 					title = title,
 					releaseDate = date,
 					order = order,
-					readingPosition = yPosition,
+					readingPosition = yPosition.toDouble(),
 					readingStatus = readChapter,
 					bookmarked = bookmarked
 				)
