+++ android/src/main/java/app/shosetsu/android/ui/reader/types/model/MarkdownReader.kt
@@ -58,7 +58,7 @@ class MarkdownReader(itemView: View) : ReaderChapterViewHolder(itemView) {
 		TODO("Not yet implemented")
 	}
 
-	override fun setProgress(progress: Int) {
+	override fun setProgress(progress: Double) {
 		TODO("Not yet implemented")
 	}
